Fixes a missed float to double change in the HAL.

+++ hal/include/HAL/HAL.h
@@ -136,7 +136,7 @@ char* HAL_GetJoystickName(int32_t joystickNum);
 int32_t HAL_GetJoystickAxisType(int32_t joystickNum, int32_t axis);
 int32_t HAL_SetJoystickOutputs(int32_t joystickNum, int64_t outputs,
                                int32_t leftRumble, int32_t rightRumble);
-float HAL_GetMatchTime(int32_t* status);
+double HAL_GetMatchTime(int32_t* status);
 
 void HAL_WaitForDSData(void);
